python-homekit/requirements.txt
paho-mqtt
python-dateutil
requests
HAP-python[QRCode]
-e libs/docker_secrets